+++ ts/packages/providers/mastra/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @composio/mastra
 
+## 0.6.2
+
+### Patch Changes
+
+- @composio/core@0.6.2
+
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies [e746383]
+  - @composio/core@0.6.1
+
 ## 0.6.0
 
 ### Minor Changes